+++ lib/src/components/snackbar_dialog.dart
@@ -1,0 +1,79 @@
+// Copyright (c) 2025. Alexandr Moroz
+
+import 'dart:async';
+import 'package:flutter/material.dart';
+import '../theme/colors.dart';
+import '../theme/constants.dart';
+import '../theme/text.dart';
+// import 'dialog.dart'; // TODO: Add dialog component
+import 'list_tile.dart';
+
+Future showMTSnackbar(String text, {TextAlign? titleAlign = TextAlign.center, Widget? trailing, Function()? onTap}) async {
+  // TODO: Implement showMTDialog
+  return;
+}
+
+class _MTSnackbarDialog extends StatefulWidget {
+  const _MTSnackbarDialog(this._text, {this.titleAlign, this.trailing, this.onTap});
+  final String _text;
+  final TextAlign? titleAlign;
+  final Widget? trailing;
+  final Function()? onTap;
+
+  @override
+  State<_MTSnackbarDialog> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<_MTSnackbarDialog> {
+  late Timer _closingTimer;
+
+  @override
+  void initState() {
+    _closingTimer = Timer(const Duration(milliseconds: 3000), () {
+      if (context.mounted) Navigator.of(context).pop();
+    });
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    _closingTimer.cancel();
+    super.dispose();
+  }
+
+  Future _onTap(BuildContext context) async {
+    _closingTimer.cancel();
+    Navigator.of(context).pop();
+    if (widget.onTap != null) {
+      widget.onTap!();
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        color: colors.b3Color.resolve(context),
+        borderRadius: BorderRadius.only(
+          topLeft: Radius.circular(constants.defBorderRadius),
+          topRight: Radius.circular(constants.defBorderRadius),
+        ),
+      ),
+      child: SafeArea(
+        child: MTListTile(
+          padding: EdgeInsets.all(constants.P3).copyWith(bottom: 0),
+          color: Colors.transparent,
+          middle: H2(
+            widget._text,
+            color: colors.f2Color,
+            align: widget.titleAlign,
+            maxLines: 3,
+          ),
+          trailing: widget.trailing,
+          bottomDivider: false,
+          onTap: () => _onTap(context),
+        ),
+      ),
+    );
+  }
+}